+++ lib/src/mixin/will_notify.dart
@@ -1,0 +1,5 @@
+import '../notifier.dart';
+
+mixin WillNotify {
+  Notifier notifier;
+}