+++ lib/sso.dart
@@ -2,5 +2,7 @@ library sso;
 
 export 'src/action_button.dart';
 export 'src/authenticator.dart';
+export 'src/mixin/will_notify.dart';
+export 'src/notifier.dart';
 export 'src/provider.dart';
 export 'src/user.dart';